IKFX-1892: misc: moto-sensorhub: Fix conversion macros.
The STM32_TO_HOST macro was casting to the wrong data type causing
values to be truncated. Both conversion macros now use the correct size
data type when casting the result value.

Reviewed-on: http://gerrit.mot.com/736322
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ include/linux/motosh.h
@@ -280,8 +280,8 @@
 
 /* The following macros are intended to be called with the stm IRQ handlers */
 /* only and refer to local variables in those functions. */
-#define STM16_TO_HOST(buf, x) ((short) be16_to_cpu(*((u16 *) (buf+(x)))))
-#define STM32_TO_HOST(buf, x) ((short) be32_to_cpu(*((u32 *) (buf+(x)))))
+#define STM16_TO_HOST(buf, x) ((int16_t) be16_to_cpu(*((u16 *) (buf+(x)))))
+#define STM32_TO_HOST(buf, x) ((int32_t) be32_to_cpu(*((u32 *) (buf+(x)))))
 
 #define MOTOSH_HALL_SOUTH 1
 #define MOTOSH_HALL_NORTH 2
